requirements.txt
asgiref==3.4.1
Django==3.2.8
djangorestframework==3.12.4
Pygments==2.10.0
httpie==2.5.0
pytz==2021.3
sqlparse==0.4.2
